+++ src/distributions/DWT1boost.cc
@@ -43,7 +43,7 @@ bool DWT1boost::checkParameterValue(vector<double const *> const &par,
   bool df_OK;
 
   // all omegas are within [0, 1] and the last omega == 1
-  for(int j = 0; j < (n_omega(len)-1); ++j){
+  for(unsigned j = 0; j < (n_omega(len)-1); ++j){
     omega_OK = omega_OK && ( omega(par)[j] >= 0.0 ) && ( omega(par)[j] <= 1.0 );
   }
   // numerical imprecission for last omega is not a problem since it's assumed to be 1 later on
@@ -77,7 +77,7 @@ double DWT1boost::logDensity(double const *x, unsigned int length, PDFType type,
   }else if(*x < crit_t(par)[0]){
     w = log(omega(par)[0]);
   }else{
-    for(int i = 1; i < n_omega(len); ++i){
+    for(unsigned i = 1; i < n_omega(len); ++i){
       if( (*x < crit_t(par)[i] ) && (*x >= crit_t(par)[i-1]) ){
         w = log(omega(par)[i]);
         break;
@@ -100,7 +100,7 @@ double DWT1boost::logDensity(double const *x, unsigned int length, PDFType type,
   denom_sum = denoms[0];
   // the ones in the middle
   if(n_omega(len) > 1){
-    for(int j = 1; j < n_omega(len) - 1; ++j){
+    for(unsigned j = 1; j < n_omega(len) - 1; ++j){
       denoms.push_back(cdf(t_dist, crit_t(par)[j]) - denom_sum);
       if(denoms[j] < 0.0){ // check and correct for possibly negative numbers due to numerical imprecission
         denoms[j] = 0.0;
@@ -115,7 +115,7 @@ double DWT1boost::logDensity(double const *x, unsigned int length, PDFType type,
   }
 
   // weight and sum the denominators
-  for(int k = 0; k < n_omega(len); ++k){
+  for(unsigned k = 0; k < n_omega(len); ++k){
     denom = denom + exp(log(denoms[k]) + log(omega(par)[k]));
   }
 
@@ -139,7 +139,7 @@ void DWT1boost::support(double *lower, double *upper, unsigned int length,
 	     vector<unsigned int> const &len) const
 {
   // no idea whether this is correct
-  for (unsigned int i = 0; i < length; ++i) {
+  for (unsigned i = 0; i < length; ++i) {
 	  lower[i] = JAGS_NEGINF;
 	  upper[i] = JAGS_POSINF;
   }